// src/game/types/UnitTypes.ts
export type UnitId =
  | "peasant"
  | "swordsman"
  | "archer"
  | "horseman"
  | "long_spearman"
  | "mace_guard"
  | "mage"
  | "knife_thrower";

export type UnitRole =
  | "economy"
  | "melee"
  | "ranged"
  | "cavalry"
  | "counter"
  | "tank"
  | "magic"
  | "support";

export type UnitVisualId =
  | "peasant"
  | "swordsman"
  | "archer"
  | "horseman"
  | "long_spearman"
  | "mace_guard"
  | "mage"
  | "knife_thrower";

export type UnitTarget = "unit" | "castle" | "resource" | "ally";